refactor(fw): Phase 4 SYS module — tear down utils/, R1.2 -Werror (Phase 4 complete)
Final Phase-4 module (7/7). Dismantles the utils/ catch-all per R2.6.

R2.6 — utils/ split by concern:
  - src/utils/utils.c → src/sys/sys_utils.c (timestamp, device-id, CRC32);
    its logging half already went to src/logger/log_text.c (LOG pass).
  - src/utils/{math_utils,timer_callbacks}.c → src/sys/.
  - include/utils/{math_utils,timer_callbacks,types}.h → include/sys/.
  - include/utils/utils.h deleted; its decls split between logger/logger.h
    (logging) and the new sys/sys_utils.h (timing/device/CRC). All include
    sites rewritten accordingly.
  Only src/utils/test_file.* remains (a standalone test/demo, not a module).

R2.1 — include/sys/sys.h umbrella over state, sys_utils, and the folded-in
math_utils / timer_callbacks / types headers.

R1.2 / CONV-06 — sys TUs flipped to full R1.2 + -Werror: heartbeat.c
(void)channel; timer_callbacks.c + sys_utils.c uint64->uint32 casts;
logger.c sign-cast (LOG pass).

Phase 4 is now complete: all 7 owned modules (ACT, EST, SNS, CTRL, COMM,
LOG, SYS) build under the full R1.2 set with -Werror on the arm target.
coding-guidelines R1.2 / R2.1 / R10.3 / CONV-06 flipped to ✅;
warning-rollout.md and requirements §2 finalised.

Verification: arm target links clean (7/7 modules -Werror); host SITL +
full 5-suite ctest green.

+++ src/sys/sys_utils.c
@@ -1,10 +1,14 @@
-/* NOTE (Phase 4): text logging (vayu_log + vayu_log_queue) moved to
- * src/logger/log_text.c (LOG module). What remains here — timestamp,
- * device-id, and CRC helpers — moves to src/sys/ in the SYS pass. */
-#include "utils/utils.h"
+/**
+ * @file src/sys/sys_utils.c
+ * @brief System utilities — timestamp, device id, CRC32.
+ *
+ * Moved here from src/utils/utils.c per Phase 4 R2.6 (the text-logging
+ * half went to src/logger/log_text.c).
+ */
+#include "sys/sys_utils.h"
+
 #include "ipc.h"
-#include "structure.h"
-#include "utils.h"
+#include "navhal.h" /* hal_crc_* */
 #include "variables.h"
 #include <stdint.h>
 
@@ -16,13 +20,14 @@ static uint8_t _device_id = 0;
 uint64_t get_timestamp(void) { return _time_stamp_high_freq; }
 
 uint32_t get_timestamp_unix(void) {
-  return _time_stamp_high_freq / (HIGH_FREQ_TIMER_FREQ / 1000) +
+  return (uint32_t)(_time_stamp_high_freq / (HIGH_FREQ_TIMER_FREQ / 1000)) +
          _time_stamp_high_freq_offset;
 }
 
 void set_timestamp(uint32_t timestamp) {
   _time_stamp_high_freq_offset =
-      timestamp - _time_stamp_high_freq / (HIGH_FREQ_TIMER_FREQ / 1000);
+      timestamp -
+      (uint32_t)(_time_stamp_high_freq / (HIGH_FREQ_TIMER_FREQ / 1000));
 }
 
 uint8_t get_device_id(void) { return _device_id; }
